crawler/requirements.txt
requests
beautifulsoup4
google-cloud-firestore
selenium
webdriver-manager 